==> notebooks/1.2-jam-wind-functions.ipynb <==
# Main script to calculate wind direction relationship from fire to patient

Modules: N/A <br>
Author: Jordan Meyer <br>
Email: jordan.meyer@berkeley.edu <br>
Date created: Feb 18, 2023 <br>

**Citations (data sources)**

``How to compute wind speed and direction:``

6. https://stackoverflow.com/questions/21484558/how-to-calculate-wind-direction-from-u-and-v-wind-components-in-r
7. https://github.com/blaylockbk/Ute_WRF/blob/master/functions/wind_calcs.py

``Wind speed and direction intuition:``

8. http://colaweb.gmu.edu/dev/clim301/lectures/wind/wind-uv
9. https://www.earthdatascience.org/courses/use-data-open-source-python/intro-vector-data-python/spatial-data-vector-shapefiles/intro-to-coordinate-reference-systems-python/

``To create maps of this wind data:``

and also used to provide intuition for winddir and windspeed

10. https://disc.gsfc.nasa.gov/information/howto?title=How%20to%20calculate%20and%20plot%20wind%20speed%20using%20MERRA-2%20wind%20component%20data%20using%20Python


**Citations (persons)**
1. Cornelia Ilin 

**Preferred environment**
1. Code written in Jupyter Notebooks

### Step 1: Import packages

import pandas as pd
import numpy as np
import netCDF4 as ncdf
import os
from datetime import date, timedelta
from math import pi
import fiona

import matplotlib.pyplot as plt
import cartopy.crs as ccrs
from cartopy.mpl.gridliner import LONGITUDE_FORMATTER, LATITUDE_FORMATTER
import matplotlib.ticker as mticker
from mpl_toolkits.axes_grid1.axes_divider import make_axes_locatable

# geography
import geopandas as gpd
import osmnx as ox
import shapely
from shapely.geometry import Point
import contextily as ctx 

#Moved from sklearn.neighbors to sklearn.metrics following their package change
import sklearn.metrics
dist = sklearn.metrics.DistanceMetric.get_metric(
    'haversine'
)

# ignore warnings
import warnings
warnings.filterwarnings(
    'ignore'
)


### Step 2: Define working directories

#Local directories on my machine (not gdrive)
in_dir_zip_shapes = '../data/raw/wind/tl_2010_06_zcta510/'
in_dir = '../data/processed/'
in_health = '../data/raw/health/'
out_dir = '../data/processed/'

# #Local directories on windows machine (not gdrive)
# in_dir_zip_shapes = 'tl_2010_06_zcta510\\'
# in_dir = 'wind\\ca only\\'
# in_health = 'health\\'
# out_dir = 'wind\\clean\\'

### Step 3: Read data

``wind``

df = pd.read_csv(in_dir + "all_years_wind_data.csv",index_col=0)

df.head()

test_df = gdf[['lat','lon','u','v','wdir','wspd','geometry']].head(10)

test_df

#### Haversine

# Radius of earth in km
r = 6378.100

def hav_dist(a,b):
    """
    Calculate the Haversine distance between two points of lat lon
    """
    
    pa_lat = a['lat']
    pa_lon = a['lon']
    pb_lat = b['lat']
    pb_lon = b['lon']

    s_lat = np.sin((np.deg2rad(pb_lat - pa_lat))/2)**2 
    s_lon = np.sin((np.deg2rad(pb_lon - pa_lon))/2)**2
    c_lat = np.cos(np.deg2rad(pb_lat)) * np.cos(np.deg2rad(pa_lat)) 
    tot = s_lat+s_lon*c_lat
    dist = 2*r*np.arcsin(np.sqrt(tot)) * 0.621371
    
    return dist

hav_dist(test_df.iloc[4,:],test_df.iloc[3,:]) 

def bearing(a,b):
    """
    Calculate the bearing from point b to point a
    """
    
    pa_lat = a['lat']
    pa_lon = a['lon']
    pb_lat = b['lat']
    pb_lon = b['lon']
    
    if pb_lon > pa_lon:
        d_X = -np.cos(pa_lat) * np.sin(pa_lon - pb_lon)
    else:
        d_X = np.cos(pa_lat) * np.sin(pa_lon - pb_lon)

    if pb_lat >pa_lat:
        d_Y = -np.cos(pb_lat) * np.sin(pa_lat) - np.sin(pb_lat) * np.cos(pa_lat) * np.cos(pa_lon - pb_lon)
    else:
        d_Y = np.cos(pb_lat) * np.sin(pa_lat) - np.sin(pb_lat) * np.cos(pa_lat) * np.cos(pa_lon - pb_lon)

    B_ba = np.arctan2(d_X,d_Y)*(180/np.pi)%360
    return B_ba
    
bearing(test_df.iloc[3,:],test_df.iloc[4,:])

def wind_direction(u,v):
    """
    Calculate direction of wind in degrees as bearing angle from true North
    """
    
    wind_dir =  np.arctan2(u,v)*(180/np.pi)%360
    return wind_dir

wind_direction(test_df.iloc[4,:]['u'],test_df.iloc[4,:]['v'])

def interaction(bearing, wdir_ba):
    """
    Calculate interaction term of wind and bearing based on difference of angles.
    """
    
    angle_diff = np.max([bearing,wdir_ba]) - np.min([bearing,wdir_ba])
    if angle_diff >180:
        upwind_effect = (angle_diff)/360
    else:
        upwind_effect = 1-(angle_diff)/360
    return upwind_effect

interaction(bearing(test_df.iloc[3,:],test_df.iloc[4,:]), wind_direction(test_df.iloc[4,:]['u'],test_df.iloc[4,:]['v']))

np.arctan2(90)*(180/np.pi)%360

upwind_effect = 1-(angle_diff)/360

import numpy as np

bearing = 270
wdir_ba = 180

angle_diff = np.max([bearing,wdir_ba]) - np.min([bearing,wdir_ba])

upwind_effect = 1-(angle_diff)/360
upwind_effect
