==> notebooks/4.0-jam-join_data.ipynb <==
# Main script to Join Data on Instrument 2 Fire Location

Modules: N/A <br>
Author: Jordan Meyer <br>
Email: jordan.meyer@berkeley.edu <br>
Date created: Feb 18, 2023 <br>

**Citations (data sources)**


**Citations (persons)**
1. Cornelia Ilin 

**Preferred environment**
1. Code written in Jupyter Notebooks

### Step 1: Import packages

import pandas as pd
import numpy as np
import netCDF4 as ncdf
import os
from datetime import date, timedelta
from math import pi
import fiona

import matplotlib.pyplot as plt
import cartopy.crs as ccrs
from cartopy.mpl.gridliner import LONGITUDE_FORMATTER, LATITUDE_FORMATTER
import matplotlib.ticker as mticker
from mpl_toolkits.axes_grid1.axes_divider import make_axes_locatable

# geography
import geopandas as gpd
import osmnx as ox
import shapely
from shapely.geometry import Point
import contextily as ctx 

#Moved from sklearn.neighbors to sklearn.metrics following their package change
import sklearn.metrics
dist = sklearn.metrics.DistanceMetric.get_metric(
    'haversine'
)

# ignore warnings
import warnings
warnings.filterwarnings(
    'ignore'
)


### Step 2: Define working directories

#Local directories on my machine (not gdrive)
in_dir = '../data/interim/'
out_dir = '../data/processed/'

ls

### Step 3: Define functions

### Step 4: Read data

## Wind

Wind load from other file

wind_df = pd.read_csv(in_dir + "all_years_wind_data.csv",index_col=0).sort_values(["year_month","ZCTA10"]).reset_index(drop=True)

wind_df.head()

## Fire

Finsh cleaning some fire things and then join fire to wind 

fire_df = pd.read_csv(in_dir + "fire_zipcodes_compressed.csv",index_col=0)

#Extract year and month, filter for fires in scope 1991+
fire_df['year'] = fire_df['ALARM_DATE'].str[:4]
fire_df = fire_df[fire_df['year'].astype(int)>1990]


#Extract Month and End Year/Month
fire_df['month'] = fire_df['ALARM_DATE'].str[5:7]
fire_df['year_month'] = fire_df['year']+fire_df['month']
fire_df['month'] = fire_df['month'].astype(int)
fire_df['end_year'] = fire_df['CONT_DATE'].str[:4]
fire_df['end_month'] = fire_df['CONT_DATE'].str[5:7].astype(int)

#Calculate duration with monthly basis, add 1 as proxy for ceiling calculation ie fire at any time in month counts regardless of start date
fire_df['duration_months'] = (fire_df['end_month']-fire_df['month'])+1
fire_df = fire_df.set_index(['ZCTA','ALARM_DATE'])

# For fires that span multiple years replace the values of the duration on a yearly basis and duplciate the entry for the next year
for idx,fire in fire_df[fire_df['end_year'] != fire_df['year']].iterrows():
    fire['FIRE_NAME'] = fire['FIRE_NAME']+" CONT"
    fire['year'] = int(fire['year'])+1
    fire['month'] = "01"
    fire['year_month'] = str(fire['year']) + fire['month']
    #Hard coding duration = 1 as all fires that spanned across years happened to be put out in January
    fire['duration_months'] = 1
    
    #Replace the total durations with the duration in year, clear up negatives durations from month 2 < month 1
    fire_df.loc[idx,'duration_months'] =  (13-fire_df.loc[idx,'month'])
    
    #Append the new year's entry for the fire's continuation
    fire_df = fire_df.append(fire)

fire_df = fire_df.reset_index()

fire_df.shape

# For all fires that are longer than 1 month insert another entry with remaing duration into the dataframe
print(fire_df[fire_df['duration_months']>1].shape)
for idx,fire in fire_df[fire_df['duration_months']>1].iterrows():
    month = 2
    duration = fire['duration_months'] 
    while duration > 1:
        # print(idx, fire_df.iloc[idx]['duration_months'], duration-1)
        #fire['FIRE_NAME'] = fire_df.iloc[idx]['FIRE_NAME'] + f"_{month}"
        fire['duration_months'] -= 1
        fire['month'] += 1
        fire['year_month'] = (str(fire['year']) + str(fire['month'])) if len(str(fire['month']))>1 else (str(fire['year']) + "0" + str(fire['month'])) 
        fire_df = fire_df.append(fire)
        duration = fire['duration_months']
        month += 1

pd_fire = pd.DataFrame(fire_df.drop('fire_centroid',axis=1))

fire_df.drop_duplicates()

pd_fire.drop_duplicates().set_index(['ZCTA','year_month'])

pd_fire['ZCTA'] = pd_fire['ZCTA'].astype(str).str[:5]
pd_fire['year_month'] = pd_fire['year_month'].str[:]


pd_fire.dtypes

fire_ready = pd_fire[~pd_fire['ZCTA'].isna()]\
    .reset_index(drop=True)\
    .drop_duplicates()


wind_ready = wind_df\
    .drop('Unnamed: 0', axis=1)\
    .rename(columns={'ZCTA10':'ZCTA'})\
    .reset_index(drop=True)\
    .drop_duplicates()

wind_ready['year_month'] = wind_ready['year_month'].astype(str)
wind_ready['ZCTA'] = wind_ready['ZCTA'].astype(str)

wind_ready.dtypes[['year_month','ZCTA']]

fire_ready.dtypes[['year_month','ZCTA']]

wind_fire = wind_ready.merge(fire_ready, on=['year_month','ZCTA'], how='left')

wind_fire

## PM2.5

Load PM2.5 from lawis files downloaded from gdrive

aqi1 = pd.read_csv(in_dir + 'finalpm25.csv', index_col = 0)
aqi1

pm25_df = aqi1.reset_index(drop=True)\
    .rename(columns={'ZIP10':'ZCTA',
                    'year_month':'old_ym'})

pm25_df.shape

pm25_df['year_month']=pm25_df['old_ym'].str[:4]+pm25_df['old_ym'].str[5:]
pm25_df['ZCTA'] = pm25_df['ZCTA'].astype(str)
pm25_ready = pm25_df.drop('old_ym',axis=1)

pm25_ready.dtypes

wfp_df = wind_fire.merge(pm25_ready, on=['year_month','ZCTA'], how='left')

wfp_df.head()

wfp_df['treatment'] = ~wfp_df['duration_months'].isna()

print(f"Number of months/zips in treatment group {np.sum(wfp_df['treatment'])}")

print(f"Number of months/zips in control group {np.sum(~wfp_df['treatment'])}")

wfp_df.dtypes

final_df = wfp_df.drop(['year_x','month_x', 'ALARM_DATE', 'FIRE_NAME', 
                        'CONT_DATE', 'CAUSE','Shape_Length', 'Shape_Area', 
                        'index_right','year_y', 'month_y','duration_months',
                        'end_year', 'end_month'],axis=1)

final_df

final_df.to_csv(out_dir + 'instrument2.csv')
