==> notebooks/7.0-jm-pd-fake_med_data.ipynb <==
# Import fire and pm2.5 data and generate fake medical data

#import relevant tools
import numpy as np
import pandas as pd
from faker import Faker
import os
from datetime import datetime
fake=Faker()

# Import fire and PM2.5 data

import sys

IN_COLAB = "google.colab" in sys.modules

if IN_COLAB:
    !pip install geopandas --quiet
    from google.colab import drive

    drive.mount("/content/drive")
    in_dir = (
        in_instrument
    ) = "/content/drive/MyDrive/capstone_fire/notebooks/instrument_2/"
    import drive.MyDrive.capstone_fire.modules.deep_ols as deep_ols
else:
    in_dir = in_instrument = "../data/instrument_2/"
    %cd '..'
    import modules.deep_ols as deep_ols
    %cd 'notebooks'

#read in data that contains fire and pm25 
fd=pd.read_csv(in_instrument+'finalish_df_6.csv') 

fd.head()

#list all files in directory
os.listdir('.')


#Take a look at the data
fd.head()


#select columns from the original data
fd1=fd[["ZCTA","year_month","zip_pm25","treatment"]]#,"DURATION","GIS_ACRES"]]

#Peek at data
fd1.head()

#check for null zipcodes
print(fd1["ZCTA"].notnull().sum())
print(fd1.shape)

zipcodes=fd1["ZCTA"]

# Creation of fake medical data
Call function by setting num = # of pa tients. 
This number(num) for california should be about 30,000,000(about 1.1M patients per year)


#create information on individual patients
#define a function to create fake medical conditions for individuals within each county
def make_patients(num):
    #list of diseases to randomly assign to patients
    disease=["PTSD","Depression","Anxiety"]
    race=["White", "Black/African American", "American Indian/Alaska Native","Asian","Native Hawaiian/Other Pacific Islander"]
    gender=["M","F"]
    
    
    fake_patients=[{'Patient Name':fake.name(),
                  'Visit_date':fake.date_between(start_date=datetime(1991,1,1),end_date=datetime(2020,12,1)),  
                  'Gender':np.random.choice(gender,p=[0.5,0.5]),
                  'Race':np.random.choice(race,p=[0.2,0.2,0.2,0.2,0.2]),
                  'Zipcode':np.random.choice(zipcodes),
                  'Patient_disease':np.random.choice(disease,p=[0.3,0.3,0.4])} for x in range(num)]

    
    return fake_patients



#call the function with the number of patients
patient_df=pd.DataFrame(make_patients(100000))
patient_df

patient_df["Visit_date1"]=patient_df["Visit_date"].astype("string")

patient_df.head()

patient_df["Visit_date_ym"]=patient_df["Visit_date1"].str[0:7]

patient_df.head()

#This is the dataset with the individual patient level information
patient_df["visit_year_month"]=patient_df["Visit_date_ym"].str.replace('-','')

patient_df.head()

#select only relevant fields  to join with fire and pm2.5 data
patient_df1=patient_df[["Patient Name","Gender","Race","Zipcode","Patient_disease","visit_year_month"]]

Aggregate individual patient level data to the zipcode level by month

#summarize the data at the zipcode level
#columns are: zipcode, month, count of patient diagnosis
patient_data_zipcode=patient_df.groupby(['Zipcode','visit_year_month','Patient_disease']).size().reset_index(name='total_cases')

patient_data_zipcode.head()

patient_data_zipcode.tail()

patient_data_zipcode.shape

# Join fire & PM2.5 data to patient medical data

fd1

#convert year_month to string from original int
fd1=fd.astype({'year_month':'string'})

#join patient and fire/pm25 datasets
#this is zipcode level data
#This is a left join(fire/pm25 is the left dataset and patient zipcode level data is the right dataset)
# The join fields are: zipcode and year_month
df=fd1.merge(patient_data_zipcode,how='left', left_on=['ZCTA','year_month'],right_on=['Zipcode','visit_year_month'] )

fd1.dtypes

patient_data_zipcode.dtypes

df = df.drop(['Zipcode','visit_year_month','Unnamed: 0'],axis=1)

df.to_csv(in_instrument + 'final_df_2.csv')

df.drop_duplicates()
